refactor: Input 디렉토리 -> component/commom 내부로 이동
